add read me and restructure the files
