+++ web/src/components/atoms/charts/GroupedBarChart.js
