+++ web/src/components/atoms/charts/StackedNormalizedHorizontalBarChart.js
